redesign and refactor
