+++ package.json
@@ -4,7 +4,8 @@
   "description": "This repo for your kanban client",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "parcel public/index.html"
   },
   "repository": {
     "type": "git",
