+++ src/main.js
@@ -5,5 +5,5 @@ import GSignInButton from 'vue-google-signin-button'
 Vue.use(GSignInButton)
 
 new Vue({
-  render: h => h(App),
+  render: h => h(App)
 }).$mount('#app');
