+++ src/views/Home/index.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+  </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
