+++ src/views/Kanban/index.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <Board />
+  </div>
+</template>
+
+<script>
+import Board from './components/Board.vue'
+export default {
+  components: {
+    Board
+  }
+}
+</script>
+
+<style>
+
+</style>
